lib/js/datachannel_test_2peer/event/004.html

<!DOCTYPE html>
<html lang="en" class="a">
	<head>
		<meta charset="utf-8">
		<!-- Testharness Timeout to long -> 60 Seconds-->
		<meta name="timeout" content="long">
		<title>WebRTC - Event Test</title>
		<script src="../resources/testharness.js"></script>
		<script src="../resources/testharnessreport.js"></script><link href="../resources/testharness.css" type="text/css" rel="stylesheet"/>
		<!-- Including the Firebase.js to send the ICE Candidate to the other peer-->
		<script type="text/javascript" src='https://cdn.firebase.com/v0/firebase.js'></script>
	</head>
	<body class="a">
		<div>
			<div id="signallingInfo"></div>
			
				
			
		</div>
		<hr />

		<!-- Testharness log -->
		<div id="log"></div>
		<script src="../datachannel2peers.js" type="text/javascript"></script>
		<script type="text/javascript">
            var label = "test";
            var dataChannel = null;
            var waitTime = 5000;
            var readyStates = "";
            var test = async_test("Set up a DataChannel - close the DataChannel (method close) - check readyState should be \"closing\" or \"closed\"");
            if (offerer) {
                test.step(function() {
                    dataChannel = peerConnection.createDataChannel(label);
                    dataChannel.onopen = test.step_func(function() {  
                    	dataChannel.close();
                    	readyStates = dataChannel.readyState; 
                        setTimeout(test.step_func(function() {
                            readyStates += dataChannel.readyState; 
                            assert_true(((readyStates.search("closed") != -1) || (readyStates.search("closing") != -1)), "ReadyState 'closing' or 'closed' not set");
                            test.done();
                        }), waitTime);
                    });
                });
            } else {
                peerConnection.ondatachannel = test.step_func(function(e) {
                    dataChannel = e.channel;
                    dataChannel.onopen = test.step_func(function() {
                    	dataChannel.close();
                    	readyStates = dataChannel.readyState; 	
                        setTimeout(test.step_func(function() {
                         	readyStates += dataChannel.readyState;    
                         	assert_true(((readyStates.search("closed") != -1) || (readyStates.search("closing") != -1)), "ReadyState 'closing' or 'closed' not set");
                         	test.done();
                        }), waitTime);
                    }); 
                });
            }
		</script>
	</body>
</html>
